requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
pydantic==1.8.2
passlib==1.7.4
python-dotenv==0.19.0
psycopg2-binary==2.9.1
